scripts/requirements.txt
azure-storage-blob==12.24.0
requests==2.32.3
